+++ builtins/src/lib.rs
@@ -24,22 +24,6 @@ fn get_globals(vm: &mut SloshVm, registers: &[Value]) -> VMResult<Value> {
     Ok(vm.alloc_vector(result))
 }
 
-fn get_globals_sorted(vm: &mut SloshVm, registers: &[Value]) -> VMResult<Value> {
-    if !registers.is_empty() {
-        return Err(VMError::new_vm(
-            "sizeof-value: takes no arguments".to_string(),
-        ));
-    }
-    let mut result = BTreeMap::new();
-    for g in vm.globals().keys() {
-        let sym = Value::Symbol(*g);
-        let val: String = sym.display_value(vm);
-        result.insert(val, sym);
-    }
-    let v = result.values().cloned().collect();
-    Ok(vm.alloc_vector(v))
-}
-
 fn get_prop(vm: &mut SloshVm, registers: &[Value]) -> VMResult<Value> {
     if registers.len() != 2 {
         return Err(VMError::new_vm(
@@ -234,17 +218,6 @@ Example:
         "Usage: (get-globals)
 
 Return a vector containing all the symbols currently defined globally.
-
-Section: core
-",
-    );
-    add_builtin(
-        env,
-        "get-globals-sorted",
-        get_globals_sorted,
-        "Usage: (get-globals-sorted)
-
-Return a vector containing all the symbols currently defined globally in sorted order (alphanumerically).
 
 Section: core
 ",
